app/requirements.txt
django>=5.0.0
djangorestframework>=3.14.0
psycopg2-binary>=2.9.0
